+++ Sources/JSONSchema.swift
@@ -21,7 +21,7 @@ public enum Type: Swift.String {
 extension String {
   func stringByRemovingPrefix(_ prefix:String) -> String? {
     if hasPrefix(prefix) {
-      let index = characters.index(startIndex, offsetBy: prefix.characters.count)
+      let index = self.index(startIndex, offsetBy: prefix.count)
       return substring(from: index)
     }
 
@@ -61,6 +61,7 @@ public struct Schema {
     formats = [
       "ipv4": validateIPv4,
       "ipv6": validateIPv6,
+      "uri": validateURI,
     ]
   }
 
@@ -224,7 +225,7 @@ func validators(_ root: Schema) -> (_ schema: [String:Any]) -> [Validator] {
           return flatten(document.map(itemsValidators))
         }
 
-        return .Valid
+        return .valid
       }
 
       validators.append(validateItems)
@@ -261,7 +262,7 @@ func validators(_ root: Schema) -> (_ schema: [String:Any]) -> [Validator] {
           return flatten(results)
         }
 
-        return .Valid
+        return .valid
       }
 
       validators.append(validateItems)
@@ -317,7 +318,7 @@ func validators(_ root: Schema) -> (_ schema: [String:Any]) -> [Validator] {
           }
         }
 
-        return .Valid
+        return .valid
       }
     }
 
@@ -329,12 +330,12 @@ func validators(_ root: Schema) -> (_ schema: [String:Any]) -> [Validator] {
               if value[dependency] == nil {
                 return .invalid(["'\(key)' is missing it's dependency of '\(dependency)'"])
               }
-              return .Valid
+              return .valid
             })
           }
         }
 
-        return .Valid
+        return .valid
       }
     }