+++ Sources/Validators.swift
@@ -11,12 +11,12 @@ import Foundation
 var isRelease = false
 
 public enum ValidationResult {
-  case Valid
+  case valid
   case invalid([String])
 
   public var valid: Bool {
     switch self {
-    case .Valid:
+    case .valid:
       return true
     case .invalid:
       return false
@@ -25,7 +25,7 @@ public enum ValidationResult {
 
   public var errors:[String]? {
     switch self {
-    case .Valid:
+    case .valid:
       return nil
     case .invalid(let errors):
       return errors
@@ -51,12 +51,12 @@ func flatten(_ results:[ValidationResult]) -> ValidationResult {
     return .invalid(errors)
   }
 
-  return .Valid
+  return .valid
 }
 
 /// Creates a Validator which always returns an valid result
 func validValidation(_ value:Any) -> ValidationResult {
-  return .Valid
+  return .valid
 }
 
 /// Creates a Validator which always returns an invalid result with the given error
@@ -75,36 +75,36 @@ func validateType(_ type: String) -> (_ value: Any) -> ValidationResult {
     case "integer":
       if let number = value as? NSNumber {
         if !CFNumberIsFloatType(number) && CFGetTypeID(number) != CFBooleanGetTypeID() {
-          return .Valid
+          return .valid
         }
       }
     case "number":
       if let number = value as? NSNumber {
         if CFGetTypeID(number) != CFBooleanGetTypeID() {
-          return .Valid
+          return .valid
         }
       }
     case "string":
       if value is String {
-        return .Valid
+        return .valid
       }
     case "object":
       if value is NSDictionary {
-        return .Valid
+        return .valid
       }
     case "array":
       if value is NSArray {
-        return .Valid
+        return .valid
       }
     case "boolean":
       if let number = value as? NSNumber {
         if CFGetTypeID(number) == CFBooleanGetTypeID() {
-          return .Valid
+          return .valid
         }
       }
     case "null":
       if value is NSNull {
-        return .Valid
+        return .valid
       }
     default:
       break
@@ -137,7 +137,7 @@ func anyOf(_ validators:[Validator], error:String? = nil) -> (_ value: Any) -> V
     for validator in validators {
       let result = validator(value)
       if result.valid {
-        return .Valid
+        return .valid
       }
     }
 
@@ -155,7 +155,7 @@ func oneOf(_ validators: [Validator]) -> (_ value: Any) -> ValidationResult {
     let validValidators = results.filter { $0.valid }.count
 
     if validValidators == 1 {
-      return .Valid
+      return .valid
     }
 
     return .invalid(["\(validValidators) validates instead `oneOf`."])
@@ -169,7 +169,7 @@ func not(_ validator: @escaping Validator) -> (_ value: Any) -> ValidationResult
       return .invalid(["'\(value)' does not match 'not' validation."])
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -182,7 +182,7 @@ func allOf(_ validators: [Validator]) -> (_ value: Any) -> ValidationResult {
 func validateEnum(_ values: [Any]) -> (_ value: Any) -> ValidationResult {
   return { value in
     if (values as! [NSObject]).contains(value as! NSObject) {
-      return .Valid
+      return .valid
     }
 
     return .invalid(["'\(value)' is not a valid enumeration value of '\(values)'"])
@@ -194,12 +194,12 @@ func validateEnum(_ values: [Any]) -> (_ value: Any) -> ValidationResult {
 func validateLength(_ comparitor: @escaping ((Int, Int) -> (Bool)), length: Int, error: String) -> (_ value: Any) -> ValidationResult {
   return { value in
     if let value = value as? String {
-      if !comparitor(value.characters.count, length) {
+      if !comparitor(value.count, length) {
         return .invalid([error])
       }
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -208,7 +208,7 @@ func validatePattern(_ pattern: String) -> (_ value: Any) -> ValidationResult {
     if let value = value as? String {
       let expression = try? NSRegularExpression(pattern: pattern, options: NSRegularExpression.Options(rawValue: 0))
       if let expression = expression {
-        let range = NSMakeRange(0, value.characters.count)
+        let range = NSMakeRange(0, value.count)
         if expression.matches(in: value, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: range).count == 0 {
           return .invalid(["'\(value)' does not match pattern: '\(pattern)'"])
         }
@@ -217,7 +217,7 @@ func validatePattern(_ pattern: String) -> (_ value: Any) -> ValidationResult {
       }
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -234,7 +234,7 @@ func validateMultipleOf(_ number: Double) -> (_ value: Any) -> ValidationResult 
       }
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -252,7 +252,7 @@ func validateNumericLength(_ length: Double, comparitor: @escaping ((Double, Dou
       }
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -266,7 +266,7 @@ func validateArrayLength(_ rhs: Int, comparitor: @escaping ((Int, Int) -> Bool),
       }
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -280,20 +280,20 @@ func validateUniqueItems(_ value: Any) -> ValidationResult {
 
     let numbers = value.filter { value in value is NSNumber } as! [NSNumber]
     let numerBooleans = numbers.filter(isBoolean)
-    let booleans = numerBooleans as [Bool]
+    let booleans = (numerBooleans as? [Bool]) ?? []
     let nonBooleans = numbers.filter { number in !isBoolean(number) }
     let hasTrueAndOne = booleans.filter { v in v }.count > 0 && nonBooleans.filter { v in v == 1 }.count > 0
     let hasFalseAndZero = booleans.filter { v in !v }.count > 0 && nonBooleans.filter { v in v == 0 }.count > 0
     let delta = (hasTrueAndOne ? 1 : 0) + (hasFalseAndZero ? 1 : 0)
 
     if (NSSet(array: value).count + delta) == value.count {
-      return .Valid
+      return .valid
     }
 
     return .invalid(["\(value) does not have unique items"])
   }
 
-  return .Valid
+  return .valid
 }
 
 // MARK: Object
@@ -306,7 +306,7 @@ func validatePropertiesLength(_ length: Int, comparitor: @escaping ((Int, Int) -
       }
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -314,13 +314,13 @@ func validateRequired(_ required: [String]) -> (_ value: Any)  -> ValidationResu
   return { value in
     if let value = value as? [String:Any] {
       if (required.filter { r in !value.keys.contains(r) }.count == 0) {
-        return .Valid
+        return .valid
       }
 
       return .invalid(["Required properties are missing '\(required)'"])
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -345,7 +345,7 @@ func validateProperties(_ properties: [String:Validator]?, patternProperties: [S
           do {
             let expression = try NSRegularExpression(pattern: pattern, options: NSRegularExpression.Options(rawValue: 0))
             let keys = value.keys.filter {
-              (key: String) in expression.matches(in: key, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: NSMakeRange(0, key.characters.count)).count > 0
+              (key: String) in expression.matches(in: key, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: NSMakeRange(0, key.count)).count > 0
             }
 
             allKeys.addObjects(from: Array(keys))
@@ -364,7 +364,7 @@ func validateProperties(_ properties: [String:Validator]?, patternProperties: [S
       return flatten(results)
     }
 
-    return .Valid
+    return .valid
   }
 }
 
@@ -401,26 +401,46 @@ func validateDependencies(_ key: String, dependencies: [String]) -> (_ value: An
 func validateIPv4(_ value:Any) -> ValidationResult {
   if let ipv4 = value as? String {
     if let expression = try? NSRegularExpression(pattern: "^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$", options: NSRegularExpression.Options(rawValue: 0)) {
-      if expression.matches(in: ipv4, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: NSMakeRange(0, ipv4.characters.count)).count == 1 {
-        return .Valid
+      if expression.matches(in: ipv4, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: NSMakeRange(0, ipv4.count)).count == 1 {
+        return .valid
       }
     }
 
     return .invalid(["'\(ipv4)' is not valid IPv4 address."])
   }
 
-  return .Valid
+  return .valid
 }
 
 func validateIPv6(_ value:Any) -> ValidationResult {
   if let ipv6 = value as? String {
     var buf = UnsafeMutablePointer<Int8>.allocate(capacity: Int(INET6_ADDRSTRLEN))
     if inet_pton(AF_INET6, ipv6, &buf) == 1 {
-      return .Valid
+      return .valid
     }
 
     return .invalid(["'\(ipv6)' is not valid IPv6 address."])
   }
 
-  return .Valid
+  return .valid
+}
+
+func validateURI(_ value:Any) -> ValidationResult {
+  if let uri = value as? String {
+    // Using the regex from http://blog.dieweltistgarnichtso.net/constructing-a-regular-expression-that-matches-uris
+
+    if let expression = try? NSRegularExpression(pattern: "((?<=\\()[A-Za-z][A-Za-z0-9\\+\\.\\-]*:([A-Za-z0-9\\.\\-_~:/\\?#\\[\\]@!\\$&'\\(\\)\\*\\+,;=]|%[A-Fa-f0-9]{2})+(?=\\)))|([A-Za-z][A-Za-z0-9\\+\\.\\-]*:([A-Za-z0-9\\.\\-_~:/\\?#\\[\\]@!\\$&'\\(\\)\\*\\+,;=]|%[A-Fa-f0-9]{2})+)", options: NSRegularExpression.Options(rawValue: 0)) {
+      let result = expression.matches(in: uri, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: NSMakeRange(0, uri.count))
+      if result.count == 1 {
+        let foundRange = result[0].range
+        if foundRange.location == 0 && foundRange.length == uri.count {
+          return .valid
+        }
+      }
+    }
+
+    return .invalid(["'\(uri)' is not a valid URI."])
+  }
+
+  return .valid
 }