+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # JSONSchema Changelog
+
+## Master
+
+### Breaking Changes
+
+- Support for Swift 3 was removed.
+- `ValidationResult.Valid` was renamed to `ValidationResult.valid`.
+
+### Enhancements
+
+- Support for Swift 4 has been added.
+- `uri` format is now validated.
+
+
 ## 0.3.0
+
 ### Enhancements
 
 - Adds support for Swift 2.2.